Improve Assets page UI

+++ client/sho-pocket-app/src/app/components/assets/asset-list/asset-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { NgForm } from '@angular/forms'
 import { ToastrService } from 'ngx-toastr';   
 
 import { AssetService } from '../../../services/asset.service';
@@ -10,18 +11,32 @@ import { Asset } from '../../../models/asset.model'
   styleUrls: ['./asset-list.component.css']
 })
 export class AssetListComponent implements OnInit {
+  currentEditRecordId: string;
+  isAddMode: boolean;
 
   constructor(public assetService : AssetService, private toastr : ToastrService) { }
 
   ngOnInit() {
+    this.assetService.getCurrenciesList();
     this.assetService.getAssetList();
   }
 
   showForEdit(asset: Asset) {
-    this.assetService.selectedAsset = Object.assign({}, asset);;
+    if (this.currentEditRecordId === asset.id) {
+      this.currentEditRecordId = null;
+    } else {
+      this.currentEditRecordId = asset.id;
+    }
   }
 
   onDelete(id: string) {
+    if (id === null)
+    {
+      this.assetService.assetList.shift();
+      this.isAddMode = false;
+      return;
+    }
+
     if (confirm('Are you sure to delete this record?') == true) {
       this.assetService.deleteAsset(id)
       .subscribe(result => {
@@ -34,5 +49,64 @@ export class AssetListComponent implements OnInit {
       })
     }
   }
-  
+
+  onSubmit(form: NgForm) {
+    if (form.value.id == null) {
+      this.assetService.postAsset(this.assetService.selectedAsset)
+        .subscribe(asset => {
+          this.toastr.success('New Record Added Succcessfully', 'Asset');
+          this.reload();
+          this.currentEditRecordId = null;
+          this.isAddMode = false;
+        });
+    }
+    else {
+      this.assetService.putAsset(form.value.id, form.value)
+      .subscribe(asset => {
+        if(asset) {
+          var listIndex = this.assetService.assetList.findIndex(a => a.id === asset.id);
+          this.assetService.assetList[listIndex] = asset;
+          this.toastr.success('Updated Successfully!', 'Asset');
+          this.currentEditRecordId = null;
+          this.isAddMode = false;
+        } else {
+          this.toastr.error('Update failed! Possibly, cannot update asset currency because balance for asset exists', 'Asset');
+        }
+      });
+    }
+  }
+
+  reload()
+  {
+    this.assetService.getAssetList();
+  }
+
+  resetRecord(id: string) {
+    this.assetService.getAsset(id).subscribe(result => {
+      let index = this.assetService.assetList.findIndex(f => f.id === id)
+      this.assetService.assetList[index] = result;
+    })
+  }
+
+  addAsset() {
+    let newAsset = {
+      id: null,
+      name: '',
+      value: 0.0,
+      currencyId: '',
+      currencyName: '',
+      isActive: true
+    }
+
+    this.assetService.assetList.unshift(newAsset);
+    this.assetService.selectedAsset = newAsset;
+    this.currentEditRecordId = null;
+    this.isAddMode = true;
+  }
+
+  onCurrencyAdded(value) {
+    let currency = this.assetService.currenciesList.find(a => a.id == value);
+    this.assetService.selectedAsset.currencyId = currency.id;
+    this.assetService.selectedAsset.currencyName = currency.name;
+  }
 }
